+++ GridKit/Model/PhasorDynamics/Converter/REPCA/Repca.cpp
@@ -1,0 +1,27 @@
+/**
+ * @file Repca.cpp
+ * @author Luke Lowery (lukel@tamu.edu)
+ * @brief Non-Enzyme instantiation for the REPCA plant-control model.
+ */
+
+#include "RepcaImpl.hpp"
+
+namespace GridKit
+{
+  namespace PhasorDynamics
+  {
+    namespace Converter
+    {
+      template <class ScalarT, typename IdxT>
+      int Repca<ScalarT, IdxT>::evaluateJacobian()
+      {
+        Log::misc() << "Evaluate Jacobian for Repca..." << std::endl;
+        Log::misc() << "Jacobian evaluation is not implemented!" << std::endl;
+        return 0;
+      }
+
+      template class Repca<double, long int>;
+      template class Repca<double, size_t>;
+    } // namespace Converter
+  } // namespace PhasorDynamics
+} // namespace GridKit