+++ GridKit/Model/PhasorDynamics/SystemModelData.hpp
@@ -12,6 +12,7 @@
 #include <GridKit/Model/PhasorDynamics/BusToSignalAdapter/BusToSignalAdapterData.hpp>
 #include <GridKit/Model/PhasorDynamics/Converter/REECB/ReecbData.hpp>
 #include <GridKit/Model/PhasorDynamics/Converter/REGCA/RegcaData.hpp>
+#include <GridKit/Model/PhasorDynamics/Converter/REPCA/RepcaData.hpp>
 #include <GridKit/Model/PhasorDynamics/Exciter/IEEET1/Ieeet1Data.hpp>
 #include <GridKit/Model/PhasorDynamics/Exciter/SEXS-PTI/SexsPtiData.hpp>
 #include <GridKit/Model/PhasorDynamics/Governor/GASTPTI/GastPtiData.hpp>
@@ -42,6 +43,7 @@ namespace GridKit
       using BusToSignalAdapterDataT = BusToSignalAdapterData<RealT, IdxT>;
       using BusFaultDataT           = BusFaultData<RealT, IdxT>;
       using RegcaDataT              = Converter::RegcaData<RealT, IdxT>;
+      using RepcaDataT              = Converter::RepcaData<RealT, IdxT>;
       using ReecbDataT              = Converter::ReecbData<RealT, IdxT>;
       using Tgov1DataT              = Governor::Tgov1Data<RealT, IdxT>;
       using GastPtiDataT            = Governor::GastPtiData<RealT, IdxT>;
@@ -98,6 +100,7 @@ namespace GridKit
       std::vector<BranchDataT>             branch;       ///< Branches within the model
       std::vector<BusFaultDataT>           bus_fault;    ///< Bus faults within the model
       std::vector<RegcaDataT>              regca;        ///< REGCA converter instances within the model
+      std::vector<RepcaDataT>              repca;        ///< REPCA plant controllers within the model
       std::vector<ReecbDataT>              reecb;        ///< REECB electrical controllers within the model
       std::vector<GenrouDataT>             genrou;       ///< GENROU instances within the model
       std::vector<GensalDataT>             gensal;       ///< GENSAL instances within the model